- refactor digitalin to c++ - cleanup for some classes affected by rtos shuffling

+++ firmware/branches/cpp/core/makingthings/serial_isr.cpp
@@ -29,7 +29,7 @@ void SerialIsr_Handler( int index );
 void SerialIsr_Handler( int index )
 {
   signed portCHAR cChar;
-  long xTaskWokenByTx = false;
+  int xTaskWokenByTx = false;
   int xTaskWokenByPost = false;
   long xTaskWokenByTxThis = false;
   long xTaskWokenByPostThis = false;
